refactor: initial EditorV2 changes
These changes are intended to simplify the Editor logic by changing how pages are stored. Pages will now be stored as a list of updates, including updates on nested blocks.

This also comes with a new strategy for syncing. On page load we send the full state to the workspace canister, the canister responds with any updates that we don't have locally and we then apply those changes locally.

A few classes have been created to support these changes: CollaborativeDocument, WebSocketProvider